Optimize code to follow DRY principles
