+++ .idea/workspace.xml
@@ -4,39 +4,7 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="c95fea79-73b1-4b27-9377-7c24a90e0339" name="Changes" comment="complete">
-      <change beforePath="$PROJECT_DIR$/.idea/material_theme_project_new.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/material_theme_project_new.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/data/constants.js" beforeDir="false" afterPath="$PROJECT_DIR$/data/constants.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/index.html" beforeDir="false" afterPath="$PROJECT_DIR$/index.html" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/App.component.js" beforeDir="false" afterPath="$PROJECT_DIR$/ui/App.component.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Common/PlayAgainButton/PlayAgainButton.component.js" beforeDir="false" afterPath="$PROJECT_DIR$/ui/Common/Buttons/PlayAgainButton/PlayAgainButton.component.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/Lose.component.js" beforeDir="false" afterPath="$PROJECT_DIR$/ui/Lose/Lose.component.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/modalDecoration/Icon/loseIcon.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/modalDecoration/modalDecoration.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/modalElements/Message/Message.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/modalElements/Results/Catch/Catch.component.js" beforeDir="false" afterPath="$PROJECT_DIR$/ui/Common/Elements/Results/Catch/Catch.component.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/modalElements/Results/Catch/Result/ResultCatch.component.js" beforeDir="false" afterPath="$PROJECT_DIR$/ui/Common/Elements/Results/Catch/Result/ResultCatch.component.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/modalElements/Results/Catch/Title/Title.component.js" beforeDir="false" afterPath="$PROJECT_DIR$/ui/Common/Elements/Results/Catch/Title/Title.component.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/modalElements/Results/Miss/Miss.component.js" beforeDir="false" afterPath="$PROJECT_DIR$/ui/Common/Elements/Results/Miss/Miss.component.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/modalElements/Results/Miss/Result/ResultMiss.component.js" beforeDir="false" afterPath="$PROJECT_DIR$/ui/Common/Elements/Results/Miss/Result/ResultMiss.component.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/modalElements/Results/Miss/Title/Title.component.js" beforeDir="false" afterPath="$PROJECT_DIR$/ui/Common/Elements/Results/Miss/Title/Title.component.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/modalElements/Results/Result.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/modalElements/Title/title.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Lose/modalElements/modalElements.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Settings/PlayGameButton/PlayGameButton.component.js" beforeDir="false" afterPath="$PROJECT_DIR$/ui/Common/Buttons/PlayGameButton/PlayGameButton.component.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Win/Decoration/Decoration.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Win/Decoration/Icon/WinIcon.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Win/Elements/Elements.component.js" beforeDir="false" afterPath="$PROJECT_DIR$/ui/Win/Elements/Elements.component.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Win/Elements/Message/Message.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Win/Elements/Results/Catch/Catch.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Win/Elements/Results/Catch/Result/ResultCatch.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Win/Elements/Results/Catch/Title/Title.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Win/Elements/Results/Miss/Miss.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Win/Elements/Results/Miss/Result/ResultMiss.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Win/Elements/Results/Miss/Title/Title.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Win/Elements/Results/Results.component.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/ui/Win/Elements/Title/Title.components.js" beforeDir="false" />
+    <list default="true" id="c95fea79-73b1-4b27-9377-7c24a90e0339" name="Changes" comment="Optimize code to follow DRY principles">
       <change beforePath="$PROJECT_DIR$/ui/Win/Win.component.js" beforeDir="false" afterPath="$PROJECT_DIR$/ui/Win/Win.component.js" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
@@ -120,7 +88,7 @@
       <updated>1719078103056</updated>
       <workItem from="1719078105796" duration="14762000" />
       <workItem from="1719093175530" duration="259000" />
-      <workItem from="1719136605764" duration="8237000" />
+      <workItem from="1719136605764" duration="8610000" />
     </task>
     <task id="LOCAL-00001" summary="add Settings">
       <option name="closed" value="true" />
@@ -138,7 +106,15 @@
       <option name="project" value="LOCAL" />
       <updated>1719092414034</updated>
     </task>
-    <option name="localTasksCounter" value="3" />
+    <task id="LOCAL-00003" summary="Optimize code to follow DRY principles">
+      <option name="closed" value="true" />
+      <created>1719144968980</created>
+      <option name="number" value="00003" />
+      <option name="presentableId" value="LOCAL-00003" />
+      <option name="project" value="LOCAL" />
+      <updated>1719144968980</updated>
+    </task>
+    <option name="localTasksCounter" value="4" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -147,6 +123,7 @@
   <component name="VcsManagerConfiguration">
     <MESSAGE value="add Settings" />
     <MESSAGE value="complete" />
-    <option name="LAST_COMMIT_MESSAGE" value="complete" />
+    <MESSAGE value="Optimize code to follow DRY principles" />
+    <option name="LAST_COMMIT_MESSAGE" value="Optimize code to follow DRY principles" />
   </component>
 </project>
