+++ ui/Win/Win.component.js
@@ -1,6 +1,6 @@
-import {ElementsComponent} from "./Elements/Elements.component.js";
 import {DecorationRenderComponent} from "../Common/Decoration/Decoration.component.js";
 import {GAME_STATUSES} from "../../data/constants.js";
+import {ElementsRenderComponent} from "../Common/Elements/Elements.component.js";
 
 export function WinComponent() {
     const element = document.createElement('div')
@@ -8,7 +8,7 @@ export function WinComponent() {
 
     element.append(
         DecorationRenderComponent(GAME_STATUSES.WIN),
-        ElementsComponent(),
+        ElementsRenderComponent(GAME_STATUSES.WIN),
 
     );
